Fix static analysis and coding standard failures
- Replace in-file AbstractWorkflowTestHarness with a dedicated
  WorkflowFollowTest subclass to resolve the PSR-1 "one class per
  file" violation that broke the analysis-and-cs job
- Drop redundant is_string() guards in AbstractWorkflowTest::header()
  (headers are array<string, string>) to fix PHPStan errors
- Assign static::class to a local before the static cache write to
  fix the Psalm PropertyTypeCoercion error
- Suppress the Psalm InvalidAttribute false-positive for #[Override]
  on methods (Psalm 6.15 reports it as "cannot be used on a function")
